[Fix] 기획 수정으로 인한 기능 수정
