--- Assets/WorkSpaces/WGH/Scripts/NPC/WGH_NPCSmellTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9cddd6181c47e0948afdafa9bc1b825b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
