+++ Assets/WorkSpaces/WGH/Scripts/NPC/WGH_NPCWait.cs
@@ -27,6 +27,8 @@ public class WGH_NPCWait : INPCState
 
     public void Enter()
     {
+        // npc 탐지 영역 true
+        controller.InteractionArea.gameObject.SetActive(true);
         // 시향 노트 찾기
         controller.TestNote = GameObject.FindGameObjectWithTag("TestNote");
         WGH_InteractionNote[] testNotes = controller.TestNote.GetComponentsInChildren<WGH_InteractionNote>();
@@ -35,37 +37,58 @@ public class WGH_NPCWait : INPCState
         {
             note.Customer = controller;
         }
-        // 시향 버튼 & 시향 종료버튼 찾기
-        controller.SmellTestStartButton = controller.TestNote.transform.GetChild(5).GetChild(5).gameObject.GetComponent<Button>();
-        controller.SmellTestEndButton = controller.TestNote.transform.GetChild(5).GetChild(6).gameObject.GetComponent<Button>();
-
-        controller.SmellTestStartButton.interactable = true;
-        controller.SmellTestStartButton.onClick.AddListener(ChangeSmellState);
 
         randomNum = Random.Range(0, (int)E_BottleType.E_BottleType_MAX);
         if (PhotonNetwork.IsMasterClient)
         {
             controller.SelectOrderUINetwork(randomNum, (int)controller.PerfumeType - 1);
         }
+
+        // 상호작용 콜라이더에 시향지가 들어와서 시향지 변수에 배정될때 그 시향지의 스크립트의 이벤트에 함수를 등록하기
+        controller.InteractionArea.GetComponent<WGH_InteractArea>().OnChangedSmellStick += FindSmellStick;
     }
 
     public void OnUpdate()
     {
         
     }
-
-    /// <summary>
-    /// 시향 상태로 전환 하는 메서드
-    /// </summary>
-    public void ChangeSmellState()
-    {
-        controller.SmellTestStartButton.interactable = false;
-        controller.ChangeStateNetwork((int)E_StateType.SMELL);
-    }
     
 
     public void Exit() 
     {
-        controller.SmellTestStartButton.onClick.RemoveListener(ChangeSmellState);
+        // npc 탐지 영역 false
+        controller.InteractionArea.gameObject.SetActive(false);
+        controller.InteractionArea.GetComponent<WGH_InteractArea>().OnChangedSmellStick -= FindSmellStick;
+    }
+
+    /// <summary>
+    /// 리액션 이벤트 등록 함수
+    /// </summary>
+    private void FindSmellStick()
+    {
+        controller.SmellStick.OnBestInteract += BestEmotion;
+        controller.SmellStick.OnLikeInteract += LikeEmotion;
+        controller.SmellStick.OnQuestionInteract += QuestionEmotion;
+        controller.SmellStick.OnDespairInteract += DespairEmotion;
+    }
+
+    private void BestEmotion()
+    {
+        controller.SelectReactUINetwork((int)E_ReactUiType.BEST);
+    }
+
+    private void LikeEmotion()
+    {
+        controller.SelectReactUINetwork((int)E_ReactUiType.LIKE);
+    }
+
+    private void QuestionEmotion()
+    {
+        controller.SelectReactUINetwork((int)E_ReactUiType.QUESTION);
+    }
+
+    private void DespairEmotion()
+    {
+        controller.SelectReactUINetwork((int)E_ReactUiType.DESPAIR);
     }
 }
