--- Assets/WorkSpaces/WGH/Scripts/WGH_CashDesk.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 204fc0e724ebf56439a71e1eee348222
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
