+++ Assets/WorkSpaces/WGH/Scripts/WGH_InteractArea.cs
@@ -1,3 +1,4 @@
+using Photon.Pun;
 using System;
 using System.Collections;
 using Unity.VisualScripting;
@@ -10,9 +11,14 @@ public class WGH_InteractArea : MonoBehaviour
     
     public event Action OnChangedSmellStick;
 
+    [SerializeField, Tooltip("실패기준 횟수")] int maxCount;
+    private int curCount;
+    [SerializeField] private bool isCheck;
+
     private void Awake()
     {
         controller = GetComponentInParent<WGH_NPCController>();
+        maxCount = 2;
     }
 
     private void OnTriggerEnter(Collider other)
@@ -22,12 +28,80 @@ public class WGH_InteractArea : MonoBehaviour
             controller.SmellStick = smellStick;
             OnChangedSmellStick?.Invoke();
         }
+        if (other.gameObject.TryGetComponent(out LSY_PotionReceiver potion) && curCount < maxCount && potion.perfumeName != E_WGH_PerfumeType.NONE)
+        {
+            if (isCheck == false && potion.perfumeName == controller.PerfumeType
+                 && potion.e_BottleType == controller.BottleType)
+            {
+                // 성공하면 성공 감정표현 후 퇴장
+                StartCoroutine(PurchaseRoutine(potion));
+            }
+            else if (isCheck == false && (potion.perfumeName != controller.PerfumeType
+                 || potion.e_BottleType != controller.BottleType))
+            {
+                // 실패하면 절망 감정표현 후 실패횟수 1회 추가
+                // 실패횟수 2회 이상일 시 퇴장
+                StartCoroutine(CheckTimeDelayRoutine(potion));
+            }
+        }
     }
     private void OnTriggerExit(Collider other)
     {
         if (other.gameObject.TryGetComponent(out WGH_SmellStick smellStick))
         {
             controller.SmellStick = null;
+        }
+    }
+    IEnumerator PurchaseRoutine(LSY_PotionReceiver potion)
+    {
+        isCheck = true;
+        controller.SelectReactUINetwork((int)E_ReactUiType.BEST);
+
+        KSD_GameManager.Instance.AddFinishPlayerCount(1);
+        yield return new WaitForSeconds(2);
+        if (curCount == 0 && KSD_GameManager.Instance != null)
+        {
+            KSD_GameManager.Instance.CurrentStageInfo.StageMoney += 100;
+        }
+        else if (curCount == 1 && KSD_GameManager.Instance != null)
+        {
+            KSD_GameManager.Instance.CurrentStageInfo.StageMoney += 80;
+        }
+        Debug.Log(KSD_GameManager.Instance.CurrentStageInfo.StageMoney);
+
+        curCount = 0;
+        controller.ChangeStateNetwork((int)E_StateType.EXIT);
+        WGH_NPCCreator.Instance.isCounter = false;
+        yield return null;
+
+        if (PhotonNetwork.IsMasterClient)
+        {
+            PhotonNetwork.Destroy(potion.photonView);
+        }
+        yield return null;
+        isCheck = false;
+    }
+
+    IEnumerator CheckTimeDelayRoutine(LSY_PotionReceiver potion)
+    {
+        isCheck = true;
+        curCount++;
+        controller.SelectReactUINetwork((int)E_ReactUiType.DESPAIR);
+        yield return new WaitForSeconds(2);
+        if (curCount >= maxCount)
+        {
+            // 실패횟수가 설정된 수에 도달하면 퇴장
+            WGH_NPCCreator.Instance.isCounter = false;
+            curCount = 0;
+            controller.ChangeStateNetwork((int)E_StateType.EXIT);
+            yield return null;
+            PhotonNetwork.Destroy(potion.photonView);
+            yield return null;
+            isCheck = false;
+        }
+        else
+        {
+            isCheck = false;
         }
     }
 }
