+++ Assets/WorkSpaces/WGH/WGH_TestScene/TestDemoScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d2fa12a637a6f6a4991bd3b27bd5a148
-PrefabImporter:
+guid: 0b7d7fc2aecb02648b29477dd841fc88
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
