+++ Assets/WorkSpaces/WGH/WGH_TestScene/WGH_TestScene.unity
@@ -257,7 +257,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 8409049458570957281, guid: 18faf9affc3364841b33c1b937437ba2, type: 3}
       propertyPath: m_RootOrder
-      value: 14
+      value: 13
       objectReference: {fileID: 0}
     - target: {fileID: 8409049458570957281, guid: 18faf9affc3364841b33c1b937437ba2, type: 3}
       propertyPath: m_LocalPosition.x
@@ -396,7 +396,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 8409049458570957281, guid: 18faf9affc3364841b33c1b937437ba2, type: 3}
       propertyPath: m_RootOrder
-      value: 13
+      value: 12
       objectReference: {fileID: 0}
     - target: {fileID: 8409049458570957281, guid: 18faf9affc3364841b33c1b937437ba2, type: 3}
       propertyPath: m_LocalPosition.x
@@ -743,7 +743,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5561765741229111064, guid: cbd9aae6a9efd3e42a7c65e0099e1923, type: 3}
       propertyPath: m_RootOrder
-      value: 12
+      value: 11
       objectReference: {fileID: 0}
     - target: {fileID: 5561765741229111064, guid: cbd9aae6a9efd3e42a7c65e0099e1923, type: 3}
       propertyPath: m_LocalPosition.x
@@ -1235,63 +1235,6 @@ Transform:
   m_Father: {fileID: 1768977672}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: -180, z: 0}
---- !u!1001 &411238790056907532
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 411238789741896944, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_Name
-      value: CashDesk
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_RootOrder
-      value: 11
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 3.863
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 1.34
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -76.07412
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 411238789741896945, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: d2fa12a637a6f6a4991bd3b27bd5a148, type: 3}
 --- !u!1001 &2521882749845886934
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1517,10 +1460,6 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 3228687245970433927, guid: 3383271c945a29847aef3d4f7139d605, type: 3}
-      propertyPath: sceneViewId
-      value: 1
-      objectReference: {fileID: 0}
     - target: {fileID: 7294067672170719924, guid: 3383271c945a29847aef3d4f7139d605, type: 3}
       propertyPath: m_RootOrder
       value: 1
